[0009] [ndf-script] keep logfile
